Table A size can be larger than 128.
- Robert Vollmert

git-svn-id: http://svn.mkgmap.org.uk/mkgmap/branches/nod@723 25d90789-57f7-4ee0-8453-03a3dfeeeb22

+++ src/uk/me/parabola/imgfmt/app/net/TableA.java
@@ -27,13 +27,13 @@ public class TableA {
 	
 	private static final int ITEM_SIZE = 5;
 
-	private byte size;
+	private int size;
 	private int numberOfItems;
 
 	public void write(ImgFileWriter writer) {
 	}
 
-	public byte getSize() {
+	public int getSize() {
 		return size;
 	}
